requirements.txt
unstructured[pdf]
langchain
langchain-community
qdrant-client
streamlit
ollama
pypdf
pdf2image
pillow
numpy
sentence-transformers
